refactor: colocate picker state APIs
