+++ datetimepicker/src/commonMain/kotlin/com/kez/picker/time/TimePicker.kt
@@ -21,8 +21,6 @@ import com.kez.picker.Picker
 import com.kez.picker.PickerColors
 import com.kez.picker.PickerDefaults
 import com.kez.picker.PickerTextStyles
-import com.kez.picker.TimePickerState
-import com.kez.picker.rememberTimePickerState
 import com.kez.picker.util.HOUR12_RANGE
 import com.kez.picker.util.HOUR24_RANGE
 import com.kez.picker.util.MINUTE_RANGE
